onboarding_de_credito_teste/requirements.txt
abstra==2.2.19
requests==2.31.0
python-dotenv==1.0.0
sendgrid==6.11.0